+++ src/components/AdminPane/AddUserForm.js
@@ -43,7 +43,7 @@ const AddUserForm = () => {
 
     try {
       const response = await fetch(
-        "http://192.168.1.12:3000/api/users/register",
+        "http://192.168.1.12:8081/api/users/register",
         {
           method: "POST",
           headers: { "Content-Type": "application/json" },
