+++ src/components/AdminPane/UserTable.js
@@ -60,7 +60,7 @@ const UserTable = () => {
       await Promise.all(
         selectedUsers.map(async (userId) => {
           const response = await fetch(
-            `http://192.168.1.12:3000/api/users/delete/${userId}`,
+            `http://192.168.1.12:3001/api/users/delete/${userId}`,
             {
               method: "DELETE",
               headers: { "Content-Type": "application/json" },
