+++ src/features/product/productSaga.js
@@ -5,7 +5,7 @@ import { fetchproductssuccess, fetchproductsfailure } from '../product/productAc
 // Function to fetch the products using fetch API
 const fetchTheApi = async () => {
   try {
-    const response = await fetch('http://192.168.1.12:3000/api/products/allproducts', {
+    const response = await fetch('http://192.168.1.12:8081/api/products/allproducts', {
       method: 'GET',
       headers: {
         'Authorization': 'Bearer YOUR_ACCESS_TOKEN', // Replace with your actual token
