+++ src/pages/HomePage.js
@@ -114,7 +114,7 @@ const HomePage = () => {
         };
     
         // API call to add product to cart
-        const response = await fetch("http://192.168.1.12:3000/api/cart/add", {
+        const response = await fetch("http://192.168.1.12:8081/api/cart/add", {
           method: "POST",
           headers: {
             "Content-Type": "application/json",
