+++ src/pages/Login.js
@@ -36,7 +36,7 @@ const Login = () => {
       );
 
       if (matchingUser) {
-        const response = await fetch("http://192.168.1.12:3000/api/users/login", {
+        const response = await fetch("http://192.168.1.12:8081/api/users/login", {
           method: "POST",
           headers: { "Content-Type": "application/json" },
           body: JSON.stringify({ email, password }),
