+++ src/pages/Register.js
@@ -51,7 +51,7 @@ const Register = () => {
       };
 
       const response = await fetch(
-        "http://192.168.1.12:3000/api/users/register",
+        "http://192.168.1.12:8081/api/users/register",
         {
           method: "POST",
           headers: { "Content-Type": "application/json" },
