+++ src/pages/UserAccontDetialsPage/WishListpage.js
@@ -44,7 +44,7 @@ const WishListPage = () => {
         };
     
         // API call to add product to cart
-        const response = await fetch("http://192.168.1.12:3000/api/cart/add", {
+        const response = await fetch("http://192.168.1.12/api/cart/add", {
           method: "POST",
           headers: {
             "Content-Type": "application/json",
